LayoutTests/editing/selection/css-pseudo-element-expected.txt

To manually test select text below by mouse and use right arrow key to navigate across the quotes. The caret should move to the right of quote.

preceding text content following text
اردو وڑ پروسیسرس پر تبصر
preceding text content following text
اردو وڑ پروسیسرس پر تبصر
PASS 1 is 1
PASS 16 is 16
PASS 1 is 1
PASS 7 is 7
PASS 1 is 1
PASS 9 is 9
PASS 8 is 8
PASS 9 is 9
PASS 1 is 1
PASS 11 is 11
PASS 10 is 10
PASS 11 is 11
PASS 1 is 1
PASS 16 is 16
PASS 1 is 1
PASS 7 is 7
PASS 10 is 10
PASS 9 is 9
PASS 8 is 8
PASS 9 is 9
PASS 1 is 1
PASS 11 is 11
PASS 1 is 1
PASS 11 is 11
PASS successfullyParsed is true

TEST COMPLETE
